// src/types/types.ts
export type FormType = {
    IPN: string
    birthday: string
    city: string
    country: string
    doc: string
    email: string
    fatherName: string
    firstName: string
    issuedBy: string
    issuedDate: string
    lastName: string
    passportNum: string
    phone: string
    secretWord: string
    sex: string
    touch: string
    uniqueNumber: string
    validUntil: string
}

export type OptionType = {
    label: string
    value: string
}